reset ui state and dismiss decimals when allocating kudos balance
